Changed minus to accept floats.

+++ movie_rating/ratings/templatetags/ratings_tags.py
@@ -10,4 +10,4 @@ def strip_page(querystring):
 
 @register.simple_tag
 def minus(a, b):
-    return int(a) - int(b)
+    return float(a) - float(b)
